Fix json code blocks
I got this warning for both blocks:

`index.rst:27: WARNING: Could not lex literal_block as "JSON". Highlighting
skipped.`

The new json isn't ideal because it makes it look like the "dependencies"
setting is the only item in the entire config, but I couldn't find a better
way. I tried to include `...` in the example but that causes errors also.