Change docs build
Change default sphinxcontrib directory to where dylan-tool install the
package if there is no 'workspace', in '_packages'.

Change 'build-docs' action and copy sphinx-contrib package to the same
directory '_packages' so the documentation can be build local and in
GH.

Change default theme to 'alabaster' to avoid install 'furo' theme as
dependency.